requirements.txt
python-dotenv
packaging
requests
tqdm
numpy
pandas
Pillow
opencv-python
pyyaml
jinja2
regex
ftfy
sentencepiece
protobuf<=3.20.3
rapidfuzz

torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

paddleocr>=2.9.1
paddlepaddle
paddlenlp

mineru==2.5.4
doclayout-yolo
ultralytics==8.0.200
onnx==1.16.0
omegaconf
shapely
pyclipper
pypdfium2

pix2tex[gui]
transformers

huggingface_hub
hf_xet
rich
